added NeuralNetwork class; fixed build issues.

+++ neuron_start/Config.h
@@ -10,23 +10,23 @@ public:
     double getThreshold() const;
     double getIncreaseWeightAmount() const;
     double getDecreaseWeightAmount() const;
-    size_t getMiddleSynapseCount() const;
-    size_t getOutSynapseCount() const;
-    size_t getMiddleNeuronCount() const;
-    size_t getSensorNeuronCount() const;
-    size_t getOutNeuronCount() const;
-    size_t getLearnLimit() const;
+    int32_t getMiddleSynapseCount() const;
+    int32_t getOutSynapseCount() const;
+    int32_t getMiddleNeuronCount() const;
+    int32_t getSensorNeuronCount() const;
+    int32_t getOutNeuronCount() const;
+    int32_t getLearnLimit() const;
 
     // setters
     void setThreshold(double threshold);
     void setIncreaseWeightAmount(double increaseWeightAmount);
     void setDecreaseWeightAmount(double decreaseWeightAmount);
-    void setMiddleSynapseCount(size_t middleSynapseCount);
-    void setOutSynapseCount(size_t outSynapseCount);
-    void setMiddleNeuronCount(size_t middleNeuronCount);
-    void setSensorNeuronCount(size_t sensorNeuronCount);
-    void setOutNeuronCount(size_t outNeuronCount);
-    void setLearnLimit(size_t learnLimit);
+    void setMiddleSynapseCount(int32_t middleSynapseCount);
+    void setOutSynapseCount(int32_t outSynapseCount);
+    void setMiddleNeuronCount(int32_t middleNeuronCount);
+    void setSensorNeuronCount(int32_t sensorNeuronCount);
+    void setOutNeuronCount(int32_t outNeuronCount);
+    void setLearnLimit(int32_t learnLimit);
 
     void printConfig() const;
     const Logger& log() const;
@@ -37,12 +37,12 @@ private:
     double _threshold;
     double _increaseWeightAmount;
     double _decreaseWeightAmount;
-    size_t _middleSynapseCount;
-    size_t _outSynapseCount;
-    size_t _middleNeuronCount;
-    size_t _sensorNeuronCount;
-    size_t _outNeuronCount;
-    size_t _learnLimit;
+    int32_t _middleSynapseCount;
+    int32_t _outSynapseCount;
+    int32_t _middleNeuronCount;
+    int32_t _sensorNeuronCount;
+    int32_t _outNeuronCount;
+    int32_t _learnLimit;
 
     // etc.
     Logger _logger;
